requirements.txt
grpcio
grpcio-tools
pytest
mysql-connector-python
